first move things into the new structure
