+++ src/submission/repositories/types.ts
@@ -1,4 +1,4 @@
-import { SubmissionId, TeamId } from '../types';
+import { SubmissionId } from '../types';
 
 export interface SubmissionRepository {
     create(sub: Omit<SubmissionDTO, 'updated'>): Promise<SubmissionDTO>;
@@ -6,17 +6,6 @@ export interface SubmissionRepository {
     findAll(): Promise<SubmissionDTO[]>;
     findById(id: SubmissionId): Promise<SubmissionDTO | null>;
     update(sub: Partial<SubmissionDTO> & { id: SubmissionId }): Promise<SubmissionDTO>;
-}
-
-export interface TeamRepository {
-    findByObjectId(id: string): Promise<TeamDTO[]>;
-    create(dtoSubmission: Omit<TeamDTO, 'updated'>): Promise<TeamDTO>;
-    update(dtoTeam: Partial<TeamDTO> & { id: TeamId }): Promise<TeamDTO>;
-}
-
-export interface TeamDTO {
-    id: TeamId;
-    updated: Date;
 }
 
 export interface SubmissionDTO {
