+++ src/submission/types.ts
@@ -1,4 +1,3 @@
 import { uuidType } from 'typesafe-uuid';
 
 export class SubmissionId extends uuidType<'SubmissionId'>() {}
-export class TeamId extends uuidType<'TeamId'>() {}
