+++ src/teams/types.ts
@@ -1,0 +1,3 @@
+import { uuidType } from 'typesafe-uuid';
+
+export class TeamId extends uuidType<'TeamId'>() {}
